requirements.txt
aiohttp
bs4
dnspython
heroku3
httpcore==0.13.7
httpx==0.19.0
motor
numpy
Pillow==8.3.2
psutil
git+https://github.com/pyrogram/pyrogram.git
PySocks==1.7.1
python-dotenv==0.19.0
pymongo
python-rapidjson
requests==2.26.0
telegraph
TgCrypto==1.2.2
urllib3==1.26.7
ujson
wget

